Move `layer::Cached` to its own module
